% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boderCol_numericInput.R
\name{boderCol_numericInput}
\alias{boderCol_numericInput}
\title{boderCol_numericInput}
\usage{
boderCol_numericInput(id, color = "#ff0000", px = 2)
}
\description{
\code{boderCol_numericInput} assigns a placeholder to a numeric input
}
\details{
This function shoudl be used inside \code{UI.R} of a shiny app.
}
\examples{

}
\author{
Javier Saez Gallego
}
