% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_color_tab.R
\name{change_color_tab}
\alias{change_color_tab}
\title{change_color_tab}
\usage{
change_color_tab(tabId, subTabNumber, color)
}
\description{
\code{change_color_tab} changes the  color of the text of a tabPanel in a tabSetPanel
}
\details{
This function should be used inside \code{server.R} of a shiny app.
}
\examples{

require(shiny)
require(shinyjs)
require(devtools)
require(shinyColors)

shinyApp(
  ui = fluidPage(
    useShinyjs(),
    mainPanel(
      h4('The colors of the tabs react to the numeric inputs.'),
      tabsetPanel(
        id = "tabSetId",
        tabPanel(
          "Input must be positive",
          numericInput('numPos', label = "This number should be > 0", value = -1)
        )
      )
    )),

  server = function(input, output) {
    observeEvent(input$numPos, {
      # Change the color of the tab when numPos is negative
      if (input$numPos <= 0) {
        change_color_tab("tabSetId", 1, 'red')
      } else{
        change_color_tab("tabSetId", 1, 'black')
      }
    })
  }
)

}
\author{
Javier Saez Gallego
}
