% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/placeholder_numericInput.R
\name{placeholder_numericInput}
\alias{placeholder_numericInput}
\title{placeholder_numericInput}
\usage{
placeholder_numericInput(id, placeHolder = "")
}
\description{
\code{placeholder_numericInput} assigns a placeholder to a numeric input
}
\details{
This function shoudl be used inside \code{UI.R} of a shiny app.
}
\examples{

}
\author{
Javier Saez Gallego
}
