% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyFiddle.R
\docType{package}
\name{shinyFiddle}
\alias{shinyFiddle}
\alias{shinyFiddle-package}
\title{shinyFiddle}
\description{
This R package includes a set of functions that allow the user to modify the following details that cannnot be modified directly by the default functions:
}
\details{
The folowing functions are included:
\itemize{
\item \code{placeholder_numericInput} adds placeholder to numericInput elements
\item \code{boderCol_numericInput} changes the boder color of a numericInput
\item \code{change_color_tab} changes the color of the tab title
\item \code{change_bgcolor_tab} changes background color
}
}
\examples{
shinyFiddle_example()

}
\author{
Javier Saez Gallego <javiersaezgallego at gmail dot com>
}
