% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{shinyFiddle_example}
\alias{shinyFiddle_example}
\title{shinyFiddle_example}
\usage{
shinyFiddle_example()
}
\description{
\code{change_bgcolor_tab} changes the background color of a tabPanel in a tabSetPanel
}
\details{
This is just an shiny showcase for the functionalities of shinyFiddle package.
}
\examples{
shinyFiddle_example()

}
\author{
Javier Saez Gallego
}
